prov/util Fix UBSAN alignment warning
This is to fix a warning from the undefined sanitizer

prov/util/src/util_eq.c:46:26: runtime error: member access within misaligned
 address 0x607000010174 for type 'struct fi_eq_err_entry', which requires
8 byte alignment 0x607000010174: note: pointer points here

The warning is a result of

ofi_eq_handle_err_entry()
    size_t err_data_size = MIN(err_entry->err_data_size,


With

ofi_eq_read()
    err_entry = (struct fi_eq_err_entry *) entry->data;

And entry::data is from struct util_event, which was indeed
not 8-byte aligned.

+++ prov/util/src/util_eq.c
@@ -104,7 +104,7 @@ ssize_t ofi_eq_read(struct fid_eq *eq_fid, uint32_t *event,
 
 			ofi_eq_handle_err_entry(eq->fabric->fabric_fid.api_version,
 						flags, err_entry, buf);
-			ret = (ssize_t) entry->size;
+			ret = entry->size;
 
 			if (!(flags & FI_PEEK))
 				eq->saved_err_data = err_entry->err_data;
@@ -143,7 +143,7 @@ ssize_t ofi_eq_write(struct fid_eq *eq_fid, uint32_t event,
 	if (!entry)
 		return -FI_ENOMEM;
 
-	entry->size = (int) len;
+	entry->size = len;
 	entry->event = event;
 	entry->err = !!(flags & UTIL_FLAG_ERROR);
 	memcpy(entry->data, buf, len);
